Updated mock data flow, turning MyBB into a real model and providing sample data.
